saving long long

+++ include/node.hpp
@@ -5,12 +5,12 @@
 class Node{
     public:
         Node();
-        Node(unsigned long int a00, unsigned long int a01, unsigned long int a10, unsigned long int a11);
+        Node(unsigned long long int a00, unsigned long long int a01, unsigned long long int a10, unsigned long long int a11);
         ~Node();
         void print();
-       unsigned long int (*getMatrix())[2];
+       unsigned long long int (*getMatrix())[2];
     private:
-       unsigned long int _matrix[2][2] = {0};
+       unsigned long long int _matrix[2][2] = {0};
 
 };
