+++ src/node.cpp
@@ -11,7 +11,7 @@ Node::Node() {
     }
 }
 
-Node::Node(unsigned long int a00, unsigned long int a01, unsigned long int a10, unsigned long int a11){
+Node::Node(unsigned long long int a00, unsigned long long int a01, unsigned long long int a10, unsigned long long int a11){
     _matrix[0][0] = a00;
     _matrix[0][1] = a01;
     _matrix[1][0] = a10;
@@ -19,10 +19,9 @@ Node::Node(unsigned long int a00, unsigned long int a01, unsigned long int a10, 
 }
 
 Node::~Node() {
-    
 }
 
-unsigned long int (*Node::getMatrix())[2] {
+unsigned long long int (*Node::getMatrix())[2] {
     return _matrix;
 }
